Avoid chatbot working at fluffy commands

+++ persistence/userDao.go
@@ -6,14 +6,16 @@ import (
 	_ "github.com/go-sql-driver/mysql"
 )
 
-// userNum, userId, guildId, userName, bbadda, type
-func InsertUser(u ds.User) (*int64, error) {
+// userNum, userId, guildId, userName, userType, wantTime, wantCnt
+
+// User 추가 (기본 공부 목표 주 5회 3시간)
+func UserSave(u ds.User) (*int64, error) {
 	db := dbConn()
 	err := db.Ping()
 	var id int64
 	if err == nil {
-		sql := `INSERT INTO  user (userId, guildId, userName, bbadda, userType) 
-		VALUES(?, ?, ?, 0, 'user')`
+		sql := `INSERT INTO  user (userId, guildId, userName, userType, wantTime, wantCnt) 
+		VALUES(?, ?, ?, 'user', '180', '5')`
 
 		stmt, _ := db.Prepare(sql)
 		res, _ := stmt.Exec(u.UserId, u.GuildId, u.UserName)
@@ -25,16 +27,16 @@ func InsertUser(u ds.User) (*int64, error) {
 }
 
 // User 조회
-func SelectUserByUserNum(userNum int) (ds.User, error) {
+func UserFindByUserNum(userNum int) (ds.User, error) {
 	db := dbConn()
 	err := db.Ping()
 	var user ds.User
 	if err == nil {
-		sql := `SELECT userNum, userId, guildId, userName, bbadda, userType
+		sql := `SELECT userNum, userId, guildId, userName, userType, wantTime, wantCnt 
 		FROM user WHERE userNum=?`
 		stmt, _ := db.Prepare(sql)
 
-		err = stmt.QueryRow(userNum).Scan(&user.UserNum, &user.UserId, &user.GuildId, &user.UserName, &user.Bbadda, &user.UserType)
+		err = stmt.QueryRow(userNum).Scan(&user.UserNum, &user.UserId, &user.GuildId, &user.UserName, &user.UserType, &user.WantTime, &user.WantCnt)
 		stmt.Close()
 	}
 	defer db.Close()
@@ -47,8 +49,7 @@ func SelectUserNumByUserIdAndGuildId(userId string, guildId string) (int, error)
 	err := db.Ping()
 	var userNum int64
 	if err == nil {
-		sql := `SELECT userNum
-		FROM user WHERE userId=? AND guildId=?`
+		sql := `SELECT userNum FROM user WHERE userId=? AND guildId=?`
 		stmt, _ := db.Prepare(sql)
 		err = stmt.QueryRow(userId, guildId).Scan(&userNum)
 		stmt.Close()
@@ -58,54 +59,20 @@ func SelectUserNumByUserIdAndGuildId(userId string, guildId string) (int, error)
 	return int(userNum), err
 }
 
-func SelectBbadda(userNum int) (int, error) {
-	db := dbConn()
-	err := db.Ping()
-	var bbadda int64
-	if err == nil {
-		sql := `SELECT bbadda
-		FROM user WHERE userNum = ?`
-		stmt, _ := db.Prepare(sql)
-		err = stmt.QueryRow(userNum).Scan(&bbadda)
-		stmt.Close()
-	}
-	defer db.Close()
-	db.Close()
-	return int(bbadda), err
-}
-
-func SelectOverBbadaa(guildId string, bbaddaLimit int) ([]ds.User, error) {
+// 해당 길드의 User 리스트 조회
+func UserListFindByGuildId(guildId string) ([]ds.User, error) {
 	db := dbConn()
 	err := db.Ping()
 	var user ds.User
 	var users []ds.User
 	if err == nil {
-		sql := `select userNum, userId, guildId, userName, bbadda, userType from user where guildId = ? and bbadda > ?`
-		stmt, _ := db.Prepare(sql)
-		res, _ := stmt.Query(guildId, bbaddaLimit)
-
-		for res.Next() {
-			res.Scan(&user.UserNum, &user.UserId, &user.GuildId, &user.UserName, &user.Bbadda, &user.UserType)
-			users = append(users, user)
-		}
-		stmt.Close()
-	}
-	defer db.Close()
-	return users, err
-}
-
-func SelectUserList(guildId string) ([]ds.User, error) {
-	db := dbConn()
-	err := db.Ping()
-	var user ds.User
-	var users []ds.User
-	if err == nil {
-		sql := `select userNum, userId, guildId, userName, bbadda, userType from user where guildId = ?`
+		sql := `SELECT userNum, userId, guildId, userName, userType, wantTime, wantCnt 
+		from user where guildId = ?`
 		stmt, _ := db.Prepare(sql)
 		res, _ := stmt.Query(guildId)
 
 		for res.Next() {
-			res.Scan(&user.UserNum, &user.UserId, &user.GuildId, &user.UserName, &user.Bbadda, &user.UserType)
+			res.Scan(&user.UserNum, &user.UserId, &user.GuildId, &user.UserName, &user.UserType, &user.WantTime, &user.WantCnt)
 			users = append(users, user)
 		}
 		stmt.Close()
@@ -120,7 +87,7 @@ func SelectGuildIdList() ([]string, error) {
 	var guildId string
 	var guildIds []string
 	if err == nil {
-		sql := "select distinct guildid from user;"
+		sql := "select distinct guildid from user"
 		stmt, _ := db.Prepare(sql)
 		res, _ := stmt.Query()
 
@@ -140,31 +107,12 @@ func UpdateUserType(u ds.User) (*int64, error) {
 	err := db.Ping()
 	var no int64
 	if err == nil {
-		sql := `UPDATE user SET type = ?
+		sql := `UPDATE user SET type = ? 
 		WHERE userNum = ?`
 		stmt, _ := db.Prepare(sql)
 		res, _ := stmt.Exec(u.UserType, u.UserId)
 		// auto_increment 반환
 		no, err = res.LastInsertId()
-		stmt.Close()
-	}
-	defer db.Close()
-	return &no, err
-}
-
-// 빠따 카운트 + 1
-func UpdateBbaddaByUserNum(userNum int) (*int64, error) {
-	db := dbConn()
-	err := db.Ping()
-	var no int64
-	if err == nil {
-		sql := `UPDATE user SET bbadda = bbadda+1
-		WHERE userNum = ?`
-
-		stmt, _ := db.Prepare(sql)
-		res, _ := stmt.Exec(userNum)
-
-		no, err = res.RowsAffected()
 		stmt.Close()
 	}
 	defer db.Close()
@@ -177,8 +125,7 @@ func DeleteUser(userId string) (*int64, error) {
 	err := db.Ping()
 	var no int64
 	if err == nil {
-		sql := `DELETE FROM user
-		WHERE userId = ?`
+		sql := `DELETE FROM user WHERE userId = ?`
 
 		stmt, _ := db.Prepare(sql)
 		res, _ := stmt.Exec(userId)
@@ -210,4 +157,62 @@ func DeleteUser(userId string) (*int64, error) {
 // 	// }
 // 	stmt.Close()
 // 	return userNum, err
+// }
+
+// Todo 삭제 - 사유 : 빠따 -> 성공 회수
+// func SelectBbadda(userNum int) (int, error) {
+// 	db := dbConn()
+// 	err := db.Ping()
+// 	var bbadda int64
+// 	if err == nil {
+// 		sql := `SELECT bbadda
+// 		FROM user WHERE userNum = ?`
+// 		stmt, _ := db.Prepare(sql)
+// 		err = stmt.QueryRow(userNum).Scan(&bbadda)
+// 		stmt.Close()
+// 	}
+// 	defer db.Close()
+// 	db.Close()
+// 	return int(bbadda), err
+// }
+
+// Todo 삭제 - 사유 : 빠따 -> 성공 회수
+// func SelectOverBbadaa(guildId string, bbaddaLimit int) ([]ds.User, error) {
+// 	db := dbConn()
+// 	err := db.Ping()
+// 	var user ds.User
+// 	var users []ds.User
+// 	if err == nil {
+// 		sql := `select userNum, userId, guildId, userName, bbadda, userType from user where guildId = ? and bbadda > ?`
+// 		stmt, _ := db.Prepare(sql)
+// 		res, _ := stmt.Query(guildId, bbaddaLimit)
+
+// 		for res.Next() {
+// 			res.Scan(&user.UserNum, &user.UserId, &user.GuildId, &user.UserName, &user.Bbadda, &user.UserType)
+// 			users = append(users, user)
+// 		}
+// 		stmt.Close()
+// 	}
+// 	defer db.Close()
+// 	return users, err
+// }
+
+// Todo 삭제 - 사유 : 빠따 -> 성공 회수
+// 빠따 카운트 + 1
+// func UpdateBbaddaByUserNum(userNum int) (*int64, error) {
+// 	db := dbConn()
+// 	err := db.Ping()
+// 	var no int64
+// 	if err == nil {
+// 		sql := `UPDATE user SET bbadda = bbadda+1
+// 		WHERE userNum = ?`
+
+// 		stmt, _ := db.Prepare(sql)
+// 		res, _ := stmt.Exec(userNum)
+
+// 		no, err = res.RowsAffected()
+// 		stmt.Close()
+// 	}
+// 	defer db.Close()
+// 	return &no, err
 // }
